+++ scripts/val_stage1.sh
@@ -1,0 +1,2 @@
+cd ..
+python val_stage1.py --val_set pretrained_val.txt